started tron effekt. tokens move already... effect rendering will be done next days

+++ Ambient Control - Entities/src/org/ambientlight/scenery/rendering/programms/configuration/TronRenderingProgrammConfiguration.java
@@ -8,17 +8,18 @@ import com.thoughtworks.xstream.annotations.XStreamAlias;
 @XStreamAlias("TronRenderingProgram")
 public class TronRenderingProgrammConfiguration extends SceneryConfiguration {
 
-	int speed;
+	double speed;
 	int lightPointAmount;
 	
-	public int getSpeed() {
+	public double getSpeed() {
 		return speed;
 	}
+
 	
-	public void setSpeed(int speed) {
+	public void setSpeed(double speed) {
 		this.speed = speed;
 	}
-	
+
 	public int getLightPointAmount() {
 		return lightPointAmount;
 	}
